tesa/api/main/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pymongo==4.6.1
motor==3.3.2
websockets==12.0
python-jose==3.3.0
python-multipart==0.0.7
python-dotenv==1.0.1
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.6.1